It is starting to work now.. This is a good thing.

+++ McSntt/McSntt/DataAbstractionLayer/Sqlite/BoatSqliteDal.cs
@@ -35,7 +35,7 @@ namespace McSntt.DataAbstractionLayer.Sqlite
                         command.Parameters.Add(new SQLiteParameter("@operational", boat.Operational));
                         insertedRows += command.ExecuteNonQuery();
 
-                        boat.BoatId = (int) db.LastInsertRowId;
+                        boat.BoatId = db.LastInsertRowId;
                     }
                 }
