+++ McSntt/McSntt/Models/Lecture.cs
@@ -9,7 +9,7 @@ namespace McSntt.Models
 {
     public class Lecture
     {
-        public virtual int LectureId { get; set; }
+        public virtual long LectureId { get; set; }
         public virtual Team Team { get; set; }
         [Column(TypeName = "DateTime2")]
         public virtual DateTime DateOfLecture { get; set; }
